deletes metric events DTOs to metric events resp. metric events client
