+++ internal/dynatrace/metric_events_client.go
@@ -1,0 +1,36 @@
+package dynatrace
+
+type MetricEvent struct {
+	Metadata          MEMetadata        `json:"metadata"`
+	ID                string            `json:"id,omitempty"`
+	MetricID          string            `json:"metricId"`
+	Name              string            `json:"name"`
+	Description       string            `json:"description"`
+	AggregationType   string            `json:"aggregationType,omitempty"`
+	EventType         string            `json:"eventType"`
+	Severity          string            `json:"severity"`
+	AlertCondition    string            `json:"alertCondition"`
+	Samples           int               `json:"samples"`
+	ViolatingSamples  int               `json:"violatingSamples"`
+	DealertingSamples int               `json:"dealertingSamples"`
+	Threshold         float64           `json:"threshold"`
+	Enabled           bool              `json:"enabled"`
+	TagFilters        []METagFilter     `json:"tagFilters,omitempty"`
+	AlertingScope     []MEAlertingScope `json:"alertingScope"`
+	Unit              string            `json:"unit,omitempty"`
+}
+type MEMetadata struct {
+	ConfigurationVersions []int  `json:"configurationVersions"`
+	ClusterVersion        string `json:"clusterVersion"`
+}
+
+type METagFilter struct {
+	Context string `json:"context"`
+	Key     string `json:"key"`
+	Value   string `json:"value"`
+}
+type MEAlertingScope struct {
+	FilterType       string       `json:"filterType"`
+	TagFilter        *METagFilter `json:"tagFilter"`
+	ManagementZoneID int64        `json:"managementZoneId,omitempty"`
+}
